old-tests/webapps/XMLHttpRequest/tests/submissions/Opera/event-loadstart.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: loadstart event</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#handler-xhr-onloadstart" data-tested-assertations="/../.." />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#event-xhr-loadstart" data-tested-assertations="/../.." />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" data-tested-assertations="/following-sibling::ol/li[9]//li[2]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();
      test.step(function() {
        var client = new XMLHttpRequest();
        client.onloadstart = test.step_func(function() {
          assert_equals(client.readyState, 1);
          test.done();
        });
        client.open("GET", "resources/well-formed.xml");
        client.send(null);
        setTimeout(test.step_func(function () {
          assert_unreached("onloadstart not called after 500 ms");
        }), 500);
      });
    </script>
  </body>
</html>
